Fix unnecessary casts

+++ branches/ndisv6/ndiswrapper/driver/ndis.c
@@ -281,11 +281,11 @@ ndis_encode_setting(struct wrap_device_setting *setting,
 	switch(type) {
 	case NdisParameterInteger:
 		param->data.integer = simple_strtol(setting->value, NULL, 0);
-		TRACE2("%u", (ULONG)param->data.integer);
+		TRACE2("%u", param->data.integer);
 		break;
 	case NdisParameterHexInteger:
 		param->data.integer = simple_strtol(setting->value, NULL, 16);
-		TRACE2("%u", (ULONG)param->data.integer);
+		TRACE2("%u", param->data.integer);
 		break;
 	case NdisParameterString:
 		RtlInitAnsiString(&ansi, setting->value);
